restructured gitlab & github gitops methods location, added GetCommits method for all gitops
