+++ internal/util/GitService.go
@@ -21,20 +21,15 @@ import (
 	"context"
 	"fmt"
 	"io/ioutil"
-	http2 "net/http"
 	"net/url"
-	"path"
 	"path/filepath"
-	"strconv"
 	"time"
 
 	bean2 "github.com/devtron-labs/devtron/api/bean"
 	"github.com/devtron-labs/devtron/internal/sql/repository"
 	"github.com/go-pg/pg"
-	"github.com/google/go-github/github"
 	"github.com/xanzy/go-gitlab"
 	"go.uber.org/zap"
-	"golang.org/x/oauth2"
 	"gopkg.in/src-d/go-git.v4"
 	"gopkg.in/src-d/go-git.v4/plumbing/object"
 	"gopkg.in/src-d/go-git.v4/plumbing/transport/http"
@@ -61,6 +56,7 @@ type GitClient interface {
 	GetRepoUrl(projectName string) (repoUrl string, err error)
 	DeleteRepository(name string) error
 	CreateReadme(name, userName, userEmailId string) (string, error)
+	GetCommits(repoName, projectName string) ([]*GitCommitDto, error)
 }
 
 type GitFactory struct {
@@ -77,6 +73,12 @@ type DetailedErrorGitOpsConfigActions struct {
 	StageErrorMap    map[string]error `json:"stageErrorMap"`
 	ValidatedOn      time.Time        `json:"validatedOn"`
 	DeleteRepoFailed bool             `json:"deleteRepoFailed"`
+}
+
+type GitCommitDto struct {
+	CommitHash string    `json:"commitHash"`
+	AuthorName string    `json:"authorName"`
+	CommitTime time.Time `json:"commitTime"`
 }
 
 func (factory *GitFactory) Reload() error {
@@ -218,13 +220,6 @@ func GetGitConfig(gitOpsRepository repository.GitOpsConfigRepository) (*GitConfi
 	return cfg, err
 }
 
-type GitLabClient struct {
-	client     *gitlab.Client
-	config     *GitConfig
-	logger     *zap.SugaredLogger
-	gitService GitService
-}
-
 func NewGitOpsClient(config *GitConfig, logger *zap.SugaredLogger, gitService GitService, gitOpsConfigRepository repository.GitOpsConfigRepository) (GitClient, error) {
 	if config.GitProvider == GITLAB_PROVIDER {
 		gitLabClient, err := NewGitLabClient(config, logger, gitService)
@@ -242,255 +237,6 @@ func NewGitOpsClient(config *GitConfig, logger *zap.SugaredLogger, gitService Gi
 		logger.Errorw("no gitops config provided, gitops will not work ")
 		return nil, nil
 	}
-}
-
-func NewGitLabClient(config *GitConfig, logger *zap.SugaredLogger, gitService GitService) (GitClient, error) {
-	var gitLabClient *gitlab.Client
-	var err error
-	if len(config.GitHost) > 0 {
-		_, err = url.ParseRequestURI(config.GitHost)
-		if err != nil {
-			return nil, err
-		}
-		gitLabClient, err = gitlab.NewClient(config.GitToken, gitlab.WithBaseURL(config.GitHost))
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		gitLabClient, err = gitlab.NewClient(config.GitToken)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	gitlabGroupId := ""
-	if len(config.GitlabGroupId) > 0 {
-		if _, err := strconv.Atoi(config.GitlabGroupId); err == nil {
-			gitlabGroupId = config.GitlabGroupId
-		} else {
-			groups, res, err := gitLabClient.Groups.SearchGroup(config.GitlabGroupId)
-			if err != nil {
-				responseStatus := 0
-				if res != nil {
-					responseStatus = res.StatusCode
-				}
-				logger.Warnw("error connecting to gitlab", "status code", responseStatus, "err", err.Error())
-			}
-			logger.Debugw("gitlab groups found ", "group", groups)
-			if len(groups) == 0 {
-				logger.Warn("no matching namespace found for gitlab")
-			}
-			for _, group := range groups {
-				if config.GitlabGroupId == group.Name {
-					gitlabGroupId = strconv.Itoa(group.ID)
-				}
-			}
-		}
-	} else {
-		return nil, fmt.Errorf("no gitlab group id found")
-	}
-	if gitlabGroupId == "" {
-		return nil, fmt.Errorf("no gitlab group id found")
-	}
-	group, _, err := gitLabClient.Groups.GetGroup(gitlabGroupId, &gitlab.GetGroupOptions{})
-	if err != nil {
-		return nil, err
-	}
-	if group != nil {
-		config.GitlabGroupPath = group.FullPath
-	}
-	logger.Debugw("gitlab config", "config", config)
-	return &GitLabClient{
-		client:     gitLabClient,
-		config:     config,
-		logger:     logger,
-		gitService: gitService,
-	}, nil
-}
-
-func (impl GitLabClient) DeleteRepository(name string) error {
-	err := impl.DeleteProject(name)
-	if err != nil {
-		impl.logger.Errorw("error in deleting repo gitlab", "project", name, "err", err)
-	}
-	return err
-}
-func (impl GitLabClient) CreateRepository(name, description, userName, userEmailId string) (url string, isNew bool, detailedErrorGitOpsConfigActions DetailedErrorGitOpsConfigActions) {
-	detailedErrorGitOpsConfigActions.StageErrorMap = make(map[string]error)
-	impl.logger.Debugw("gitlab app create request ", "name", name, "description", description)
-	repoUrl, err := impl.GetRepoUrl(name)
-	if err != nil {
-		impl.logger.Errorw("error in getting repo url ", "gitlab project", name, "err", err)
-		detailedErrorGitOpsConfigActions.StageErrorMap[GetRepoUrlStage] = err
-		return "", false, detailedErrorGitOpsConfigActions
-	}
-	if len(repoUrl) > 0 {
-		detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, GetRepoUrlStage)
-		return repoUrl, false, detailedErrorGitOpsConfigActions
-	} else {
-		url, err = impl.createProject(name, description)
-		if err != nil {
-			detailedErrorGitOpsConfigActions.StageErrorMap[CreateRepoStage] = err
-			return "", true, detailedErrorGitOpsConfigActions
-		}
-		detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, CreateRepoStage)
-	}
-	repoUrl = url
-	validated, err := impl.ensureProjectAvailability(name)
-	if err != nil {
-		impl.logger.Errorw("error in ensuring project availability ", "gitlab project", name, "err", err)
-		detailedErrorGitOpsConfigActions.StageErrorMap[CloneHttpStage] = err
-		return "", true, detailedErrorGitOpsConfigActions
-	}
-	if !validated {
-		detailedErrorGitOpsConfigActions.StageErrorMap[CloneHttpStage] = fmt.Errorf("unable to validate project:%s in given time", name)
-		return "", true, detailedErrorGitOpsConfigActions
-	}
-	detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, CloneHttpStage)
-	_, err = impl.CreateReadme(fmt.Sprintf("%s/%s", impl.config.GitlabGroupPath, name), userName, userEmailId)
-	if err != nil {
-		impl.logger.Errorw("error in creating readme ", "gitlab project", name, "err", err)
-		detailedErrorGitOpsConfigActions.StageErrorMap[CreateReadmeStage] = err
-		return "", true, detailedErrorGitOpsConfigActions
-	}
-	detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, CreateReadmeStage)
-	validated, err = impl.ensureProjectAvailabilityOnSsh(name, repoUrl)
-	if err != nil {
-		impl.logger.Errorw("error in ensuring project availability ", "gitlab project", name, "err", err)
-		detailedErrorGitOpsConfigActions.StageErrorMap[CloneSshStage] = err
-		return "", true, detailedErrorGitOpsConfigActions
-	}
-	if !validated {
-		detailedErrorGitOpsConfigActions.StageErrorMap[CloneSshStage] = fmt.Errorf("unable to validate project:%s in given time", name)
-		return "", true, detailedErrorGitOpsConfigActions
-	}
-	detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, CloneSshStage)
-	return url, true, detailedErrorGitOpsConfigActions
-}
-
-func (impl GitLabClient) DeleteProject(projectName string) (err error) {
-	impl.logger.Infow("deleting project ", "gitlab project name", projectName)
-	_, err = impl.client.Projects.DeleteProject(fmt.Sprintf("%s/%s", impl.config.GitlabGroupPath, projectName))
-	return err
-}
-func (impl GitLabClient) createProject(name, description string) (url string, err error) {
-	var namespace = impl.config.GitlabGroupId
-	namespaceId, err := strconv.Atoi(namespace)
-	if err != nil {
-		return "", err
-	}
-
-	// Create new project
-	p := &gitlab.CreateProjectOptions{
-		Name:                 gitlab.String(name),
-		Description:          gitlab.String(description),
-		MergeRequestsEnabled: gitlab.Bool(true),
-		SnippetsEnabled:      gitlab.Bool(false),
-		Visibility:           gitlab.Visibility(gitlab.PrivateVisibility),
-		NamespaceID:          &namespaceId,
-	}
-	project, _, err := impl.client.Projects.CreateProject(p)
-	if err != nil {
-		impl.logger.Errorw("err in creating gitlab app", "req", p, "name", name, "err", err)
-		return "", err
-	}
-	impl.logger.Infow("gitlab app created", "name", name, "url", project.HTTPURLToRepo)
-	return project.HTTPURLToRepo, nil
-}
-
-func (impl GitLabClient) ensureProjectAvailability(projectName string) (bool, error) {
-	pid := fmt.Sprintf("%s/%s", impl.config.GitlabGroupPath, projectName)
-	count := 0
-	verified := false
-	for count < 3 && !verified {
-		count = count + 1
-		_, res, err := impl.client.Projects.GetProject(pid, &gitlab.GetProjectOptions{})
-		if err != nil {
-			return verified, err
-		}
-		if res.StatusCode >= 200 && res.StatusCode <= 299 {
-			verified = true
-			return verified, nil
-		}
-		time.Sleep(10 * time.Second)
-	}
-	return false, nil
-}
-
-func (impl GitLabClient) ensureProjectAvailabilityOnSsh(projectName string, repoUrl string) (bool, error) {
-	count := 0
-	for count < 3 {
-		count = count + 1
-		_, err := impl.gitService.Clone(repoUrl, fmt.Sprintf("/ensure-clone/%s", projectName))
-		if err == nil {
-			impl.logger.Infow("gitlab ensureProjectAvailability clone passed", "try count", count, "repoUrl", repoUrl)
-			return true, nil
-		}
-		if err != nil {
-			impl.logger.Errorw("gitlab ensureProjectAvailability clone failed", "try count", count, "err", err)
-		}
-		time.Sleep(10 * time.Second)
-	}
-	return false, nil
-}
-
-func (impl GitLabClient) GetRepoUrl(projectName string) (repoUrl string, err error) {
-	pid := fmt.Sprintf("%s/%s", impl.config.GitlabGroupPath, projectName)
-	prop, res, err := impl.client.Projects.GetProject(pid, &gitlab.GetProjectOptions{})
-	if err != nil {
-		impl.logger.Debugw("gitlab get project err", "pid", pid, "err", err)
-		if res != nil && res.StatusCode == 404 {
-			return "", nil
-		}
-		return "", err
-	}
-	if res.StatusCode >= 200 && res.StatusCode <= 299 {
-		return prop.HTTPURLToRepo, nil
-	}
-	return "", nil
-}
-
-func (impl GitLabClient) CreateReadme(name, userName, userEmailId string) (string, error) {
-	fileAction := gitlab.FileCreate
-	filePath := "README.md"
-	fileContent := "devtron licence"
-	actions := &gitlab.CreateCommitOptions{
-		Branch:        gitlab.String("master"),
-		CommitMessage: gitlab.String("test commit"),
-		Actions:       []*gitlab.CommitActionOptions{{Action: &fileAction, FilePath: &filePath, Content: &fileContent}},
-		AuthorEmail:   &userEmailId,
-		AuthorName:    &userName,
-	}
-	c, _, err := impl.client.Commits.CreateCommit(name, actions)
-	return c.ID, err
-}
-func (impl GitLabClient) checkIfFileExists(projectName, ref, file string) (exists bool, err error) {
-	_, _, err = impl.client.RepositoryFiles.GetFileMetaData(fmt.Sprintf("%s/%s", impl.config.GitlabGroupPath, projectName), file, &gitlab.GetFileMetaDataOptions{Ref: &ref})
-	return err == nil, err
-}
-
-func (impl GitLabClient) CommitValues(config *ChartConfig) (commitHash string, err error) {
-	branch := "master"
-	path := filepath.Join(config.ChartLocation, config.FileName)
-	exists, err := impl.checkIfFileExists(config.ChartRepoName, branch, path)
-	var fileAction gitlab.FileActionValue
-	if exists {
-		fileAction = gitlab.FileUpdate
-	} else {
-		fileAction = gitlab.FileCreate
-	}
-	actions := &gitlab.CreateCommitOptions{
-		Branch:        &branch,
-		CommitMessage: gitlab.String(config.ReleaseMessage),
-		Actions:       []*gitlab.CommitActionOptions{{Action: &fileAction, FilePath: &path, Content: &config.FileContent}},
-		AuthorEmail:   &config.UserEmailId,
-		AuthorName:    &config.UserName,
-	}
-	c, _, err := impl.client.Commits.CreateCommit(fmt.Sprintf("%s/%s", impl.config.GitlabGroupPath, config.ChartRepoName), actions)
-	if err != nil {
-		return "", err
-	}
-	return c.ID, err
 }
 
 type ChartConfig struct {
@@ -638,245 +384,4 @@ func (impl GitServiceImpl) Pull(repoRoot string) (err error) {
 		return nil
 	}
 	return err
-}
-
-//github
-
-type GitHubClient struct {
-	client *github.Client
-
-	//config *GitConfig
-	logger                 *zap.SugaredLogger
-	org                    string
-	gitService             GitService
-	gitOpsConfigRepository repository.GitOpsConfigRepository
-}
-
-func NewGithubClient(host string, token string, org string, logger *zap.SugaredLogger,
-	gitService GitService, gitOpsConfigRepository repository.GitOpsConfigRepository) (GitHubClient, error) {
-	ctx := context.Background()
-	httpTransport := &http2.Transport{}
-	httpClient := &http2.Client{Transport: httpTransport}
-	ts := oauth2.StaticTokenSource(
-		&oauth2.Token{AccessToken: token},
-	)
-	ctx = context.WithValue(ctx, oauth2.HTTPClient, httpClient)
-
-	tc := oauth2.NewClient(ctx, ts)
-	var client *github.Client
-	var err error
-	hostUrl, err := url.Parse(host)
-	if err != nil {
-		logger.Errorw("error in creating git client ", "host", hostUrl, "err", err)
-		return GitHubClient{}, err
-	}
-	if hostUrl.Host == GITHUB_HOST {
-		client = github.NewClient(tc)
-	} else {
-		logger.Infow("creating github EnterpriseClient with org", "host", host, "org", org)
-		hostUrl.Path = path.Join(hostUrl.Path, GITHUB_API_V3)
-		client, err = github.NewEnterpriseClient(hostUrl.String(), hostUrl.String(), tc)
-	}
-
-	return GitHubClient{
-		client:                 client,
-		org:                    org,
-		logger:                 logger,
-		gitService:             gitService,
-		gitOpsConfigRepository: gitOpsConfigRepository,
-	}, err
-}
-func (impl GitHubClient) DeleteRepository(name string) error {
-	gitOpsConfig, err := impl.gitOpsConfigRepository.GetGitOpsConfigByProvider(GITHUB_PROVIDER)
-	if err != nil {
-		if err == pg.ErrNoRows {
-			gitOpsConfig.GitHubOrgId = ""
-		} else {
-			impl.logger.Errorw("error in fetching gitOps github config", "err", err)
-			return err
-		}
-	}
-	_, err = impl.client.Repositories.Delete(context.Background(), gitOpsConfig.GitHubOrgId, name)
-	if err != nil {
-		impl.logger.Errorw("repo deletion failed for github", "repo", name, "err", err)
-		return err
-	}
-	return nil
-}
-func (impl GitHubClient) CreateRepository(name, description, userName, userEmailId string) (url string, isNew bool, detailedErrorGitOpsConfigActions DetailedErrorGitOpsConfigActions) {
-	detailedErrorGitOpsConfigActions.StageErrorMap = make(map[string]error)
-	ctx := context.Background()
-	repoExists := true
-	url, err := impl.GetRepoUrl(name)
-	if err != nil {
-		responseErr, ok := err.(*github.ErrorResponse)
-		if !ok || responseErr.Response.StatusCode != 404 {
-			impl.logger.Errorw("error in creating github repo", "err", err)
-			detailedErrorGitOpsConfigActions.StageErrorMap[GetRepoUrlStage] = err
-			return "", false, detailedErrorGitOpsConfigActions
-		} else {
-			repoExists = false
-		}
-	}
-	if repoExists {
-		detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, GetRepoUrlStage)
-		return url, false, detailedErrorGitOpsConfigActions
-	}
-	private := true
-	//	visibility := "private"
-	r, _, err := impl.client.Repositories.Create(ctx, impl.org,
-		&github.Repository{Name: &name,
-			Description: &description,
-			Private:     &private,
-			//			Visibility:  &visibility,
-		})
-	if err != nil {
-		impl.logger.Errorw("error in creating github repo, ", "repo", name, "err", err)
-		detailedErrorGitOpsConfigActions.StageErrorMap[CreateRepoStage] = err
-		return "", true, detailedErrorGitOpsConfigActions
-	}
-	logger.Infow("github repo created ", "r", r.CloneURL)
-	detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, CreateRepoStage)
-
-	validated, err := impl.ensureProjectAvailabilityOnHttp(name)
-	if err != nil {
-		impl.logger.Errorw("error in ensuring project availability github", "project", name, "err", err)
-		detailedErrorGitOpsConfigActions.StageErrorMap[CloneHttpStage] = err
-		return *r.CloneURL, true, detailedErrorGitOpsConfigActions
-	}
-	if !validated {
-		detailedErrorGitOpsConfigActions.StageErrorMap[CloneHttpStage] = fmt.Errorf("unable to validate project:%s in given time", name)
-		return "", true, detailedErrorGitOpsConfigActions
-	}
-	detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, CloneHttpStage)
-
-	_, err = impl.CreateReadme(name, userName, userEmailId)
-	if err != nil {
-		impl.logger.Errorw("error in creating readme github", "project", name, "err", err)
-		detailedErrorGitOpsConfigActions.StageErrorMap[CreateReadmeStage] = err
-		return *r.CloneURL, true, detailedErrorGitOpsConfigActions
-	}
-	detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, CreateReadmeStage)
-
-	validated, err = impl.ensureProjectAvailabilityOnSsh(name, *r.CloneURL)
-	if err != nil {
-		impl.logger.Errorw("error in ensuring project availability github", "project", name, "err", err)
-		detailedErrorGitOpsConfigActions.StageErrorMap[CloneSshStage] = err
-		return *r.CloneURL, true, detailedErrorGitOpsConfigActions
-	}
-	if !validated {
-		detailedErrorGitOpsConfigActions.StageErrorMap[CloneSshStage] = fmt.Errorf("unable to validate project:%s in given time", name)
-		return "", true, detailedErrorGitOpsConfigActions
-	}
-	detailedErrorGitOpsConfigActions.SuccessfulStages = append(detailedErrorGitOpsConfigActions.SuccessfulStages, CloneSshStage)
-	//_, err = impl.createReadme(name)
-	return *r.CloneURL, true, detailedErrorGitOpsConfigActions
-}
-
-func (impl GitHubClient) CreateReadme(repoName, userName, userEmailId string) (string, error) {
-	cfg := &ChartConfig{
-		ChartName:      repoName,
-		ChartLocation:  "",
-		FileName:       "README.md",
-		FileContent:    "@devtron",
-		ReleaseMessage: "readme",
-		ChartRepoName:  repoName,
-		UserName:       userName,
-		UserEmailId:    userEmailId,
-	}
-	hash, err := impl.CommitValues(cfg)
-	if err != nil {
-		impl.logger.Errorw("error in creating readme github", "repo", repoName, "err", err)
-	}
-	return hash, err
-}
-
-func (impl GitHubClient) CommitValues(config *ChartConfig) (commitHash string, err error) {
-	branch := "master"
-	path := filepath.Join(config.ChartLocation, config.FileName)
-	ctx := context.Background()
-	newFile := false
-	fc, _, _, err := impl.client.Repositories.GetContents(ctx, impl.org, config.ChartRepoName, path, &github.RepositoryContentGetOptions{Ref: branch})
-	if err != nil {
-		responseErr, ok := err.(*github.ErrorResponse)
-		if !ok || responseErr.Response.StatusCode != 404 {
-			impl.logger.Errorw("error in creating repo github", "err", err, "config", config)
-			return "", err
-		} else {
-			newFile = true
-		}
-	}
-	currentSHA := ""
-	if !newFile {
-		currentSHA = *fc.SHA
-	}
-	timeNow := time.Now()
-	options := &github.RepositoryContentFileOptions{
-		Message: &config.ReleaseMessage,
-		Content: []byte(config.FileContent),
-		SHA:     &currentSHA,
-		Branch:  &branch,
-		Author: &github.CommitAuthor{
-			Date:  &timeNow,
-			Email: &config.UserEmailId,
-			Name:  &config.UserName,
-		},
-		Committer: &github.CommitAuthor{
-			Date:  &timeNow,
-			Email: &config.UserEmailId,
-			Name:  &config.UserName,
-		},
-	}
-	c, _, err := impl.client.Repositories.CreateFile(ctx, impl.org, config.ChartRepoName, path, options)
-	if err != nil {
-		impl.logger.Errorw("error in commit github", "err", err, "config", config)
-		return "", err
-	}
-	return *c.SHA, nil
-}
-
-func (impl GitHubClient) GetRepoUrl(projectName string) (repoUrl string, err error) {
-	ctx := context.Background()
-	repo, _, err := impl.client.Repositories.Get(ctx, impl.org, projectName)
-	if err != nil {
-		return "", err
-	}
-	return *repo.CloneURL, nil
-}
-
-func (impl GitHubClient) ensureProjectAvailabilityOnHttp(projectName string) (bool, error) {
-	count := 0
-	for count < 3 {
-		count = count + 1
-		_, err := impl.GetRepoUrl(projectName)
-		if err == nil {
-			return true, nil
-		}
-		responseErr, ok := err.(*github.ErrorResponse)
-		if !ok || responseErr.Response.StatusCode != 404 {
-			impl.logger.Errorw("error in validating repo github", "project", projectName, "err", err)
-			return false, err
-		} else {
-			impl.logger.Errorw("error in validating repo github", "project", projectName, "err", err)
-		}
-		time.Sleep(10 * time.Second)
-	}
-	return false, nil
-}
-
-func (impl GitHubClient) ensureProjectAvailabilityOnSsh(projectName string, repoUrl string) (bool, error) {
-	count := 0
-	for count < 3 {
-		count = count + 1
-		_, err := impl.gitService.Clone(repoUrl, fmt.Sprintf("/ensure-clone/%s", projectName))
-		if err == nil {
-			impl.logger.Infow("github ensureProjectAvailability clone passed", "try count", count, "repoUrl", repoUrl)
-			return true, nil
-		}
-		if err != nil {
-			impl.logger.Errorw("github ensureProjectAvailability clone failed", "try count", count, "err", err)
-		}
-		time.Sleep(10 * time.Second)
-	}
-	return false, nil
 }
